feat(home page): add header
