+++ app/hooks/use-search-stocks.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { useQuery } from "@tanstack/react-query";
-import { searchStocks } from "../queries/search-stock";
+import { searchStocks } from "../queries/search-stocks";
 
 export function useSearchStocks(searchParams: string) {
   return useQuery({
